pyproject.toml
[project]
name = "ecovision"
version = "0.1.0"
description = """
Modelo de Inteligencia Artificial para la detección y segmentación de latas 
y botellas como residuos contaminantes del medio ambiente.
"""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "dvc[azure]",
  "streamlit>=1.28.0",
  "torch>=2.0.0",
  "ultralytics>=8.0.20",
  "opencv-python>=4.8.0",
  "numpy>=1.24.0",
  "pillow>=10.0.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "mlflow>=3.4.0",
  "azure-storage-blob",
  "pytest>=8.0.0"
]
